feat(warehouse): add update warehouse command and constructor.

+++ StockSip.Platform.API/InventoryManagement/Domain/Model/Commands/UpdateWarehouseCommand.cs
@@ -13,4 +13,4 @@ namespace StockSip.Platform.API.InventoryManagement.Domain.Model.Commands;
 /// <param name="Temperature">The temperature range in the warehouse</param>
 /// <param name="ImageUrl">The image url of the warehouse</param>
 /// <param name="ProfileId">The unique identifier of the warehouse</param>
-public record UpdateWarehouseCommand(long WarehouseId, string Name, WarehouseAddress Address, Capacity Capacity, Temperature Temperature, string ImageUrl, long ProfileId);
+public record UpdateWarehouseCommand(int WarehouseId, string Name, WarehouseAddress Address, Capacity Capacity, Temperature Temperature, string ImageUrl, int ProfileId);
